+++ packages/pdf-viewer/src/plugins/PDFContextPrintPlugin/PdfContextPrintPlugin.tsx
@@ -1,18 +1,23 @@
 import React, { useContext } from 'react';
 import printJS from 'print-js';
+import { PrinterOutlined } from '@ant-design/icons';
 import PDFContextMenuPlugin from '../PDFContextMenuPlugin';
 import PDFViewerContext from '../../context';
+import { useLocale } from '../../locale/context';
+import zhCN from '../../locale/zh_CN';
 
 export interface PdfContextPrintPluginProps {}
 
 const PdfContextPrintPlugin = (props: PdfContextPrintPluginProps) => {
   const { pdfViewer } = useContext(PDFViewerContext);
+  const [locale] = useLocale(zhCN);
   return (
     <PDFContextMenuPlugin
       menu={[
         {
           key: 'print',
-          text: '打印',
+          text: locale.printCurrentFile || '打印当前文件',
+          icon: <PrinterOutlined />,
           onClick: async () => {
             const data = await pdfViewer.getPDFInstance()?.getData();
             if (data) {