+++ packages/pocket/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orca-fe/pocket
 
+## 3.3.1
+
+### Patch Changes
+
+- chore: pdf-viewer locale
+
 ## 3.3.0
 
 ### Minor Changes